refactor: 영화 데이터를 추상화하여 Presenter에 주입
