+++ app/src/main/java/woowacourse/movie/data/DummyData.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.common
+package woowacourse.movie.data
 
 import androidx.annotation.DrawableRes
 import woowacourse.movie.R
