+++ app/src/main/java/woowacourse/movie/view/movie/model/AdUiModel.kt
@@ -3,7 +3,7 @@ package woowacourse.movie.view.movie.model
 import android.os.Parcelable
 import androidx.annotation.DrawableRes
 import kotlinx.parcelize.Parcelize
-import woowacourse.movie.common.DummyData
+import woowacourse.movie.data.DummyData
 
 @Parcelize
 data class AdUiModel(
